requirements.txt
python-dotenv

pandas
xlrd
openpyxl

mysql-connector-python

streamlit
watchdog
